+++ geo/src/algorithm/bool_ops/mod.rs
@@ -2,23 +2,25 @@ mod i_overlay_integration;
 #[cfg(test)]
 mod tests;
 
-use crate::bool_ops::i_overlay_integration::convert::{
-    multi_polygon_from_shapes, ring_to_shape_path,
-};
-use crate::bool_ops::i_overlay_integration::BoolOpsCoord;
-use i_overlay::core::fill_rule::FillRule;
-use i_overlay::float::clip::FloatClip;
-use i_overlay::float::single::SingleFloatOverlay;
-use i_overlay::string::clip::ClipRule;
+use i_overlay_integration::convert::{multi_polygon_from_shapes, ring_to_shape_path};
+use i_overlay_integration::BoolOpsCoord;
 pub use i_overlay_integration::BoolOpsNum;
 
 use crate::geometry::{LineString, MultiLineString, MultiPolygon, Polygon};
+use crate::winding_order::{Winding, WindingOrder};
+
+use i_overlay::core::fill_rule::FillRule;
+use i_overlay::core::overlay_rule::OverlayRule;
+use i_overlay::float::clip::FloatClip;
+use i_overlay::float::overlay::FloatOverlay;
+use i_overlay::float::single::SingleFloatOverlay;
+use i_overlay::string::clip::ClipRule;
 
 /// Boolean Operations on geometry.
 ///
 /// Boolean operations are set operations on geometries considered as a subset
-/// of the 2-D plane. The operations supported are: intersection, union, xor or
-/// symmetric difference, and set-difference on pairs of 2-D geometries and
+/// of the 2-D plane. The operations supported are: intersection, union,
+/// symmetric difference (xor), and set-difference on pairs of 2-D geometries and
 /// clipping a 1-D geometry with self.
 ///
 /// These operations are implemented on [`Polygon`] and the [`MultiPolygon`]
@@ -34,6 +36,11 @@ use crate::geometry::{LineString, MultiLineString, MultiPolygon, Polygon};
 /// In particular, taking `union` with an empty geom should remove degeneracies
 /// and fix invalid polygons as long the interior-exterior requirement above is
 /// satisfied.
+///
+/// # Performance
+///
+/// For union operations on a large number of [`Polygon`]s or [`MultiPolygons`],
+/// using [`unary_union`] will yield far better performance.
 pub trait BooleanOps {
     type Scalar: BoolOpsNum;
 
@@ -57,18 +64,26 @@ pub trait BooleanOps {
         multi_polygon_from_shapes(shapes)
     }
 
+    /// Returns the overlapping regions shared by both `self` and `other`.
     fn intersection(
         &self,
         other: &impl BooleanOps<Scalar = Self::Scalar>,
     ) -> MultiPolygon<Self::Scalar> {
         self.boolean_op(other, OpType::Intersection)
     }
+
+    /// Combines the regions of both `self` and `other` into a single geometry, removing
+    /// overlaps and merging boundaries.
     fn union(&self, other: &impl BooleanOps<Scalar = Self::Scalar>) -> MultiPolygon<Self::Scalar> {
         self.boolean_op(other, OpType::Union)
     }
+
+    /// The regions that are in either `self` or `other`, but not in both.
     fn xor(&self, other: &impl BooleanOps<Scalar = Self::Scalar>) -> MultiPolygon<Self::Scalar> {
         self.boolean_op(other, OpType::Xor)
     }
+
+    /// The regions of `self` which are not in `other`.
     fn difference(
         &self,
         other: &impl BooleanOps<Scalar = Self::Scalar>,
@@ -107,6 +122,75 @@ pub enum OpType {
     Union,
     Difference,
     Xor,
+}
+
+/// Efficient [union](BooleanOps::union) of many adjacent / overlapping geometries
+///
+/// This is typically much faster than `union`ing a bunch of geometries together one at a time.
+///
+/// Note: Geometries can be wound in either direction, but the winding order must be consistent,
+/// and the polygon's interiors must be wound opposite to its exterior.
+///
+/// See [Orient] for more information.
+///
+/// [Orient]: crate::algorithm::orient::Orient
+///
+/// # Arguments
+///
+/// `boppables`: A collection of `Polygon` or `MultiPolygons` to union together.
+///
+/// returns the union of all the inputs.
+///
+/// # Examples
+///
+/// ```
+/// use geo::algorithm::unary_union;
+/// use geo::wkt;
+///
+/// let right_piece = wkt!(POLYGON((4. 0.,4. 4.,8. 4.,8. 0.,4. 0.)));
+/// let left_piece = wkt!(POLYGON((0. 0.,0. 4.,4. 4.,4. 0.,0. 0.)));
+///
+/// // touches neither right nor left piece
+/// let separate_piece = wkt!(POLYGON((14. 10.,14. 14.,18. 14.,18. 10.,14. 10.)));
+///
+/// let polygons = vec![left_piece, separate_piece, right_piece];
+/// let actual_output = unary_union(&polygons);
+///
+/// let expected_output = wkt!(MULTIPOLYGON(
+///     // left and right piece have been combined
+///     ((0. 0., 0. 4., 8. 4., 8. 0.,  0. 0.)),
+///     // separate piece remains separate
+///     ((14. 10., 14. 14., 18. 14.,18. 10., 14. 10.))
+/// ));
+/// assert_eq!(actual_output, expected_output);
+/// ```
+pub fn unary_union<'a, B: BooleanOps + 'a>(
+    boppables: impl IntoIterator<Item = &'a B>,
+) -> MultiPolygon<B::Scalar> {
+    let mut winding_order: Option<WindingOrder> = None;
+    let subject = boppables
+        .into_iter()
+        .flat_map(|boppable| {
+            let rings = boppable.rings();
+            rings
+                .map(|ring| {
+                    if winding_order.is_none() {
+                        winding_order = ring.winding_order();
+                    }
+                    ring_to_shape_path(ring)
+                })
+                .collect::<Vec<_>>()
+        })
+        .collect::<Vec<_>>();
+
+    let fill_rule = if winding_order == Some(WindingOrder::Clockwise) {
+        FillRule::Positive
+    } else {
+        FillRule::Negative
+    };
+
+    let shapes = FloatOverlay::with_subj(&subject).overlay(OverlayRule::Subject, fill_rule);
+    multi_polygon_from_shapes(shapes)
 }
 
 impl<T: BoolOpsNum> BooleanOps for Polygon<T> {